requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
openai==1.30.1
pinecone-client==3.2.2
langchain==0.2.3
langchain-openai==0.1.8
langchain-community==0.2.3
weasyprint==62.3
python-dotenv==1.0.1
pyyaml==6.0.1
jinja2==3.1.4
python-multipart==0.0.9
sse-starlette==2.1.0
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
